% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.dof.R
\name{calc_df_mars}
\alias{calc_df_mars}
\title{Calculate degrees of freedom of MARS}
\usage{
calc_df_mars(n = 100, p = 10, N = 100, nk = 5, d = 1, penalty = 2, seed = 1234)
}
\arguments{
\item{n}{sample size}

\item{p}{number of features}

\item{N}{number of repetitions}

\item{nk}{maximum number of model terms before pruning in earth::earth}

\item{d}{maximum degree of interactions}

\item{penalty}{penalty per knot}

\item{seed}{random seed for construct x and y}
}
\value{
a vector of two elements: \describe{
\item{dfcov}{empirical df based on covariance formula}
\item{dfapp}{average df based on mars' approximation formula}
}
}
\description{
Calculate degrees of freedom of MARS
}
