% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.dof.R
\name{correct_df}
\alias{correct_df}
\title{Correct df given a fitted earth.object}
\usage{
correct_df(x, N = 100)
}
\arguments{
\item{x}{an earth object}

\item{N}{number of repetitions for approximating empirical df}
}
\value{
a list of three elements \describe{
\item{df_cov}{empirical df based on covariance formula}
\item{df_app}{average df based on mars' approximation formula}
\item{penalty}{penalty factor}
}
}
\description{
Correct df given a fitted earth.object
}
