% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag.esl.trainset}
\alias{flag.esl.trainset}
\title{Datasets
The indicator for training and test sets for the Spam email data. See also: \url{https://hastie.su.domains/ElemStatLearn/}}
\format{
## `flag.esl.trainset`
}
\usage{
flag.esl.trainset
}
\description{
Datasets
The indicator for training and test sets for the Spam email data. See also: \url{https://hastie.su.domains/ElemStatLearn/}
}
\keyword{datasets}
