% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.dof.R
\name{sol_mars_df_and_penalty}
\alias{sol_mars_df_and_penalty}
\title{Solve penalty factor by equating empirical df}
\usage{
sol_mars_df_and_penalty(n = 20, p = 2, N = 100, nk = 5, d = 1, penalty = NULL)
}
\arguments{
\item{n}{sample size}

\item{p}{number of features}

\item{N}{number of repetitions}

\item{nk}{maximum number of model terms before pruning in earth::earth}

\item{d}{maximum degree of interactions}

\item{penalty}{the initial penalty. If not specified, it is `d + 1`.}
}
\value{
a list of three elements \describe{
\item{df_cov}{empirical df based on covariance formula}
\item{df_app}{average df based on mars' approximation formula}
\item{penalty}{penalty factor}
}
}
\description{
Solve penalty factor by equating empirical df
}
\note{
Note that `penalty` only works in the backward procedure, so we do not
need to a while loop to ensure the self-consistency.
}
