% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spam}
\alias{spam}
\title{Datasets
Spam email data. Also accessible from \url{https://archive.ics.uci.edu/dataset/94/spambase}, \url{https://hastie.su.domains/ElemStatLearn/}, or \url{https://github.com/szcf-weiya/ESL-CN/tree/master/data/Spam}.}
\format{
## `spam`
}
\usage{
spam
}
\description{
Datasets
Spam email data. Also accessible from \url{https://archive.ics.uci.edu/dataset/94/spambase}, \url{https://hastie.su.domains/ElemStatLearn/}, or \url{https://github.com/szcf-weiya/ESL-CN/tree/master/data/Spam}.
}
\keyword{datasets}
