% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.dof.R
\name{sol_mars_df_and_penalty}
\alias{sol_mars_df_and_penalty}
\title{Solve penalty factor by equating empirical df}
\usage{
sol_mars_df_and_penalty(
  n = 20,
  p = 2,
  N = 100,
  nk = 5,
  tol = 0.001,
  maxiter = 10,
  verbose = FALSE,
  fixxy = TRUE,
  d = 1
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of features}

\item{N}{number of repetitions}

\item{nk}{maximum number of model terms before pruning in earth::earth}

\item{tol}{stop criteria}

\item{maxiter}{maximum number of iterations}

\item{verbose}{whether to print debug information}

\item{fixxy}{whether to fix x and y}

\item{d}{maximum degree of interactions}
}
\value{
a list of three elements \describe{
\item{df_cov}{empirical df based on covariance formula}
\item{df_app}{average df based on mars' approximation formula}
\item{penalty}{penalty factor}
}
}
\description{
Solve penalty factor by equating empirical df
}
